ml/requirements.txt
torch==2.1.2
torchvision==0.16.2
coremltools==7.0
onnx==1.15.0
onnxruntime==1.17.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
wandb==0.16.1
PyYAML==6.0.1
albumentations==1.3.1
opencv-python-headless==4.8.1.78
timm==0.9.12
httpx==0.25.2
